adjusts

+++ requests.go
@@ -2,11 +2,11 @@ package main
 
 type sendRequest struct {
 	MessageType string `json:"type"`
-	Msg         uint64 `json:"msg"`
+	Msg         int    `json:"msg"`
 	Key         string `json:"key"`
 }
 
 type sendResponse struct {
 	MessageType string `json:"type"`
-	Offset      uint64 `json:"offset"`
+	Offset      int    `json:"offset"`
 }
